+++ packages/backend/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 1.0.0-beta.37
+
+### Patch Changes
+
+- Updated dependencies [[`bab2e7e05`](https://github.com/clerk/javascript/commit/bab2e7e0590d0da1fd7db0680e63e8f2eb836b41)]:
+  - @clerk/shared@2.0.0-beta.23
+
 ## 1.0.0-beta.36
 
 ### Minor Changes