+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 5.0.0-beta.40
+
+### Patch Changes
+
+- Updated dependencies [[`bab2e7e05`](https://github.com/clerk/javascript/commit/bab2e7e0590d0da1fd7db0680e63e8f2eb836b41)]:
+  - @clerk/shared@2.0.0-beta.23
+  - @clerk/types@4.0.0-beta.29
+
 ## 5.0.0-beta.39
 
 ### Minor Changes